requirements.txt
# Enhanced PyTorch with CUDA support for attention mechanisms
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Advanced data processing and visualization
numpy>=1.21.0
Pillow>=8.0.0
matplotlib>=3.5.0
opencv-python>=4.5.0
scipy>=1.7.0

# 3D model processing and mesh generation
trimesh>=3.15.0
open3d>=0.17.0  # Alternative 3D processing
pyvista>=0.38.0  # 3D visualization

# Computer vision and image processing
scikit-image>=0.19.0
imageio>=2.9.0
albumentations>=1.3.0  # Advanced image augmentation

# Machine learning utilities
scikit-learn>=1.0.0
pandas>=1.3.0

# Utility libraries
webcolors>=1.12.0
tqdm>=4.62.0  # Progress bars
seaborn>=0.11.0  # Enhanced plotting
